pyproject.toml
[tool.poetry]
name = "python-formatter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
pandas = "^1.4.3"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
